+++ src/Hover.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <Geode/Geode.hpp>
-#include "LimitedCCMenu.hpp"
 
 using namespace geode::prelude;